Implement product creation

+++ server/internal/gql/resolvers/userprofile.go
@@ -46,7 +46,7 @@ func (r *queryResolver) UserProfile(ctx context.Context, email *string) (*gqlMod
 	if email != nil && *email != ""{
 		userToFind = actions.GetUserWithEmail(*email, r.ORM.DB.New())
 	} else {
-		userToFind = actions.GetUserWithId(authLevel.UID, r.ORM.DB.New())
+		userToFind = actions.GetUserWithId(uint(authLevel.UID), r.ORM.DB.New())
 	}
 
 	if userToFind == nil {
